- OSX warning fixes (fixes #772) - Coverity scan fixes

+++ prov/sockets/src/sock_comm.c
@@ -125,7 +125,7 @@ ssize_t sock_comm_send(struct sock_conn *conn, const void *buf, size_t len)
 
 ssize_t sock_comm_recv_socket(struct sock_conn *conn, void *buf, size_t len)
 {
-	int ret;
+	ssize_t ret;
 	
 	ret = read(conn->sock_fd, buf, len);
 	if (ret < 0) {
